% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggSentiment.R
\name{aggSentiment}
\alias{aggSentiment}
\title{Helper function to aggregate sentiment variables
This is just a simple helper function that is used to aggregate the
sentiment variables}
\usage{
aggSentiment(inputData, speakerId)
}
\arguments{
\item{inputData}{data.table}

\item{speakerId}{character}
}
\value{
list
}
\description{
Helper function to aggregate sentiment variables
This is just a simple helper function that is used to aggregate the
sentiment variables
}
\examples{
agg.out = aggSentiment(inputData=sample_transcript_sentiment_processed, speakerId="userName")
}
