% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchFlattenSelf.R
\name{batchFlattenSelf}
\alias{batchFlattenSelf}
\title{Wrapper to run the flattenSelf function on a batch of meetings}
\usage{
batchFlattenSelf(inputData, inputType, meetingId, speakerId, gapLength)
}
\arguments{
\item{inputData}{data.frame with multiple meetings from batchProcessZoomOutput (either transcript or chat)}

\item{inputType}{character indicating 'transcript' or 'chat'}

\item{meetingId}{character name of the variable containing the meeting identifier}

\item{speakerId}{character name of the variable containing the speaker identifier}

\item{gapLength}{integer giving the number of seconds for marking distinct turns by the same speaker. Consecutive utterances by the same speaker of greater than or equal to this value will be treated as different conversational turns.}
}
\value{
a data.frame that is the same format as inputData, but where the observations are the new, compressed conversational turns.
}
\description{
Wrapper to run the flattenSelf function on a batch of meetings
}
\examples{
newChat = batchFlattenSelf(inputData = sample_chat_processed, 
inputType="chat", meetingId = "batchMeetingId", 
speakerId="userName", gapLength=120)

newTranscript = batchFlattenSelf(inputData = sample_transcript_processed, 
inputType="transcript", meetingId = "batchMeetingId", 
speakerId="userName", gapLength=120)
}
