% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchProcessZoomOutput.R
\name{batchProcessZoomOutput}
\alias{batchProcessZoomOutput}
\title{Batch process files that have been downloaded from Zoom}
\usage{
batchProcessZoomOutput(batchInput)
}
\arguments{
\item{batchInput}{String giving the location of the xlsx file
that contains the information for the zoom meetings}
}
\value{
a list that has a list item for each of the elements
of a Zoom output that are available--batchInfo, meetInfo, partInfo,
transcript, chat, and rosetta
}
\description{
Provide the location of a structured batchInput file and this
function will process a set of meetings at once.
}
\examples{
\dontrun{
batchOut = batchProcessZoomOutput(batchInput="myMeetingsBatch.xlsx")
}
}
