% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convDate}
\alias{convDate}
\title{Date Calculator from Date}
\usage{
convDate(oldDate)
}
\arguments{
\item{oldDate}{character date string}
}
\value{
character new format for date
}
\description{
transforms the Zoom output for a timestamp into YYYY-MM-DD HH:MM:SS
}
\keyword{internal}
