% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVideoDateTimeMeta.R
\name{getVideoDateTimeMeta}
\alias{getVideoDateTimeMeta}
\title{Extracts the date/time that a video was created}
\usage{
getVideoDateTimeMeta(inputVideo, tz)
}
\arguments{
\item{inputVideo}{path to a video file}

\item{tz}{the timezone code for the video file}
}
\value{
a character object in YYYY-MM-DD HH:MM:SS that is the date/time that the video was created
}
\description{
Using ffmpeg, this function extracts the creation timestamp for a given video. It is important to note that this is in UTC - so, you must have some knowledge of the timestamp to sync up with other data streams
}
\examples{
vidCreate = getVideoDateTimeMeta(inputVideo=
system.file('extdata', "meeting001_video.mp4", 
package = 'zoomGroupStats'), tz="America/Chicago")
\dontrun{
vidCreate = getVideoDateTimeMeta(inputVideo=
"myVideo.mp4", tz="America/Chicago")
}
}
