% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabVideoStills.R
\name{grabVideoStills}
\alias{grabVideoStills}
\title{Helper function to split a video into still frames}
\usage{
grabVideoStills(inputVideo, sampleWindow, tryffmpeg = TRUE)
}
\arguments{
\item{inputVideo}{full filepath to a video file}

\item{sampleWindow}{an integer indicating how frequently you want to sample
images in number of seconds.}

\item{tryffmpeg}{boolean indicating whether you want it to try to use ffmpeg,
which tends to be much faster than magick for splitting frames}
}
\value{
a data.frame that gives information about the still frames. Each record is
a stillframe, with the following info:
\itemize{
\item imageSeconds - number of seconds from the start of the video when this image was captured
\item imageName - full path to where the image has been saved as a .png
}
}
\description{
This function currently relies on either magick or
ffmpeg to split a video file into images. This function will save
the images locally in the same directory where the video is located.
}
\examples{
\dontrun{
grabVideoStills(inputVideo='myMeeting.mp4', 
sampleWindow=45)
}
}
