% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageDetectTimeStamp.R
\name{imageDetectTimeStamp}
\alias{imageDetectTimeStamp}
\title{Detects the timestamp within a recorded Zoom session}
\usage{
imageDetectTimeStamp(imageInput)
}
\arguments{
\item{imageInput}{the full path to the image file}
}
\value{
returns the timestamp as character YYYY-MM-DD HH:MM:SS
}
\description{
This function uses the 'AWS Rekognition' service to detect the timestamp within a Zoom recording. This presumes that the user included the timestamp option in their Zoom Cloud Recording setup.
}
\examples{
\dontrun{
timeStamp = imageDetectTimeStamp(imageInput="./imageFile.png")
}
}
