% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importZoomRosetta.R
\name{importZoomRosetta}
\alias{importZoomRosetta}
\title{Helper function to add unique identifiers to processed Zoom downloads
Import an edited zoomRosetta file that tells how to
link up Zoom display names to some unique individual
identifier}
\usage{
importZoomRosetta(zoomOutput, zoomRosetta, meetingId)
}
\arguments{
\item{zoomOutput}{the output of batchProcessZoomOutput}

\item{zoomRosetta}{the path to an edited zoomRosetta xlsx}

\item{meetingId}{the name of the meetingId you want to use}
}
\value{
returns zoomOutput with identifiers in zoomRosetta
merged to any available data.frames in the zoomOutput file
}
\description{
Helper function to add unique identifiers to processed Zoom downloads
Import an edited zoomRosetta file that tells how to
link up Zoom display names to some unique individual
identifier
}
\examples{
\dontrun{
importZoomRosetta(zoomOutput=batchOut, zoomRosetta="myEditedRosetta.xlsx", 
meetingId="batchMeetingId")
}
}
