% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTimeWindows.R
\name{makeTimeWindows}
\alias{makeTimeWindows}
\title{Helper function that creates temporal windows in datasets}
\usage{
makeTimeWindows(inputData, timeVar, windowSize)
}
\arguments{
\item{inputData}{data.frame that is output from processZoomTranscript or processZoomChat}

\item{timeVar}{name of a numeric column that contains the time variable you want to use}

\item{windowSize}{numeric value giving the length of time window}
}
\value{
list with two data.frames. The first is inputData with the temporal windows.
The second is the full set of temporal windows.
}
\description{
Helper function that creates temporal windows in datasets
}
\examples{
win.out = makeTimeWindows(sample_transcript_processed, 
timeVar="utteranceStartSeconds", windowSize=10)
}
