% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processZoomAudio.R
\name{processZoomAudio}
\alias{processZoomAudio}
\title{Zoom Audio File Processing}
\usage{
processZoomAudio(
  bucketName,
  jobName,
  localDir,
  speakerNames = c(),
  recordingStartDateTime,
  writeTranscript = TRUE
)
}
\arguments{
\item{bucketName}{character    name of the s3 bucket where the finished transcript is stored}

\item{jobName}{character    name of the transcription job}

\item{localDir}{character a local directory where you can save the aws json file and also a plain text file of the transcribed text}

\item{speakerNames}{character vector The Zoom user names of the speakers, in the order in which they appear in the audio clip.}

\item{recordingStartDateTime}{character the date/time that the meeting recording started $MIKE$ format?}

\item{writeTranscript}{boolean Do you want to output a plain text file of the transcript? Default is TRUE}
}
\value{
A list of analyses
   \item{utterance_id}{an incremented numeric identifier for a marked speech utterance}
   \item{utterance_start_seconds}{the number of seconds from the start of the recording (when it starts)}
   \item{utterance_start_time}{the timestamp for the start of the utterance}
   \item{utterance_end_seconds}{the number of seconds from the start of the recording (when it ends)}
   \item{utterance_end_time}{the timestamp for the end of the utterance}
   \item{utterance_time_window}{the number of seconds that the utterance took}
   \item{user_name}{the name attached to the utterance}
   \item{utterance_message}{the text of the utterance}
   \item{utterance_language}{the language code for the transcript}
}
\description{
This function parses transcriptions from AWS transcribe.
}
\details{
This function parses the JSON transcription completed by AWS transcribe. The output is the same as the processZoomTranscript function.
}
\examples{

\dontrun{
audio.out = processZoomAudio(bucketName = "my-transcription-bucket",
jobName = "mylocalfile.m4a",
localDir = "path-to-local-directory-for-output",
speakerNames = c("Tom Smith", "Jamal Jones", "Jamika Jensen"),
recordingStartDateTime = "2020-06-20 17:00:00",
writeTranscript=TRUE)
}

}
