% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processZoomChat.R
\name{processZoomChat}
\alias{processZoomChat}
\title{Process a Zoom chat file}
\usage{
processZoomChat(
  fname,
  sessionStartDateTime = "1970-01-01 00:00:00",
  languageCode = "en"
)
}
\arguments{
\item{fname}{String that is the path to the exported Zoom .txt chat file}

\item{sessionStartDateTime}{String that is the start of the session in YYYY-MM-DD HH:MM:SS}

\item{languageCode}{String denoting the language}
}
\value{
data.frame where each record is a message submission in the chat.
}
\description{
This function parses the data from the chatfile that is downloaded from the Zoom website.
NOTE: This is the file that accompanies a recording. This is not the file that you download directly within the window.
It is also not the one that is saved locally on your computer.
This is the file that you can access after a session if you record in the cloud.
}
\examples{
ch.out = processZoomChat(
fname=system.file('extdata', "sample_chat.txt", package = 'zoomGroupStats'), 
sessionStartDateTime = '2020-04-20 13:30:00', 
languageCode = 'en')
}
