% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processZoomOutput.R
\name{processZoomOutput}
\alias{processZoomOutput}
\title{Wrapper function to process the raw files from Zoom in a single call
The user provides a fileRoot that is used for a given meeting. Output
files should be named as fileRoot_chat.txt; fileRoot_transcript.vtt;
and fileRoot_participants.csv. Any relevant files will be processed.}
\usage{
processZoomOutput(
  fileRoot,
  rosetta = TRUE,
  sessionStartDateTime = "1970-01-01 00:00:00",
  recordingStartDateTime = "1970-01-01 00:00:00",
  languageCode = "en"
)
}
\arguments{
\item{fileRoot}{string giving the path to the files and the root}

\item{rosetta}{boolean to produce the rosetta file or not}

\item{sessionStartDateTime}{string giving the start of the session in YYYY-MM-DD HH:MM:SS}

\item{recordingStartDateTime}{string giving the start of the recording in YYYY-MM-DD HH:MM:SS}

\item{languageCode}{string giving the language code}
}
\value{
a named list containing data.frames for each of the available files
(particiapnts, transcript, chat)
}
\description{
Wrapper function to process the raw files from Zoom in a single call
The user provides a fileRoot that is used for a given meeting. Output
files should be named as fileRoot_chat.txt; fileRoot_transcript.vtt;
and fileRoot_participants.csv. Any relevant files will be processed.
}
\examples{
\dontrun{
zoomOut = processZoomOutput(fileRoot="~/zoomMeetings/myMeeting", rosetta=TRUE)
}
}
