% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processZoomTranscript.R
\name{processZoomTranscript}
\alias{processZoomTranscript}
\title{Process Zoom transcript file}
\usage{
processZoomTranscript(fname, recordingStartDateTime, languageCode)
}
\arguments{
\item{fname}{charracter}

\item{recordingStartDateTime}{character}

\item{languageCode}{character}
}
\value{
data.frame
}
\description{
Process Zoom transcript file
}
\examples{
tr.out = processZoomTranscript(
fname=system.file('extdata', "sample_transcript.vtt", package = 'zoomGroupStats'),
recordingStartDateTime = '2020-04-20 13:30:00', 
languageCode = 'en')
}
