% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_batch_info}
\alias{sample_batch_info}
\title{Parsed batch info file in a recorded 'Zoom' meeting}
\format{
A data frame with 3 rows of 13 variables:
\describe{
\item{batchMeetingId}{a character meeting identification variable}
\item{fileRoot}{the prefix to the files for this particular meeting}
\item{participants}{binary indicating whether there is a participants file downloaded}
\item{transcript}{binary indicating whether there is a transcript file downloaded}
\item{chat}{binary indicating whether there is a chat file downloaded}
\item{video}{binary indicating whether there is a video file downloaded}
\item{sessionStartDateTime}{start of the actual session as a character YYYY-MM-DD HH:MM:SS}
\item{recordingStartDateTime}{start of the actual recording as a character YYYY-MM-DD HH:MM:SS}
\item{participants_processed}{binary indicating whether there is a participants file already processed}
\item{transcript_processed}{binary indicating whether there is a transcript file already processed}
\item{chat_processed}{binary indicating whether there is a chat file already processed}
\item{video_processed}{binary indicating whether there is a video file already processed}
\item{dirRoot}{character giving the directory in which all files will be found}
}
}
\source{
\url{http://zoomgroupstats.org/}
}
\usage{
sample_batch_info
}
\description{
Parsed batch info file in a recorded 'Zoom' meeting
}
\keyword{datasets}
