% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_chat_processed}
\alias{sample_chat_processed}
\title{Parsed chat file in a Zoom meeting}
\format{
A data frame with 10 rows of 6 variables:
\describe{
\item{messageId}{an incremented numeric identifier for a marked chat message}
\item{messageSeconds}{when the message was posted as the number of seconds from the start of the recording}
\item{messageTime}{timestamp for message}
\item{userName}{Zoom display name attached to the messager}
\item{message}{text of the message}
\item{messageLanguage}{language code of the message}
}
}
\source{
\url{http://apknight.org/}
}
\usage{
sample_chat_processed
}
\description{
Parsed chat file in a Zoom meeting
}
\keyword{datasets}
