% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_chat_sentiment_syu}
\alias{sample_chat_sentiment_syu}
\title{Parsed chat file in a 'Zoom' meeting with sentiment analysis using syuzhet}
\format{
A data frame with 30 rows of 30 variables:
\describe{
\item{batchMeetingId}{a character meeting identification variable}
\item{messageId}{an incremented numeric identifier for a marked chat message}
\item{userName}{'Zoom' display name attached to the messager}
\item{messageSeconds}{when the message was posted as the number of seconds from the start of the recording}
\item{messageTime}{timestamp for message}
\item{message}{text of the message}
\item{messageLanguage}{language code of the message}
\item{userEmail}{character email address}
\item{userId}{numeric id of each speaker}
\item{wordCount}{number of words in this utterance}
\item{syu_anger}{number of anger words}
\item{syu_anticipation}{number of anticipation words}
\item{syu_disgust}{number of disgust words}
\item{syu_fear}{number of fear words}
\item{syu_joy}{number of joy words}
\item{syu_sadness}{number of sadness words}
\item{syu_surprise}{number of surprise words}
\item{syu_trust}{number of trust words}
\item{syu_negative}{number of negative words}
\item{syu_positive}{number of positive words}
}
}
\source{
\url{http://zoomgroupstats.org/}
}
\usage{
sample_chat_sentiment_syu
}
\description{
Parsed chat file in a 'Zoom' meeting with sentiment analysis using syuzhet
}
\keyword{datasets}
