% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textSentiment.R
\name{textSentiment}
\alias{textSentiment}
\title{Conduct a sentiment analysis on text data}
\usage{
textSentiment(inputData, idVar, textVar, languageCodeVar)
}
\arguments{
\item{inputData}{data.frame}

\item{idVar}{character}

\item{textVar}{character}

\item{languageCodeVar}{character}
}
\value{
data.frame
}
\description{
This function takes in the output of the chat and transcript functions. It then
conducts a sentiment analysis on an identified chunk of text
and returns the values.
To use this function, you must have an aws account that with privileges for the comprehend service
However you authenticate for AWS, you should do so before running the function.
}
\examples{
sent.out = textSentiment(inputData=sample_transcript_processed, idVar='userName', 
textVar='utteranceMessage', languageCodeVar='utteranceLanguage')
}
