% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timeCalc}
\alias{timeCalc}
\title{Time Calculator}
\usage{
timeCalc(incTime)
}
\arguments{
\item{incTime}{character time string}
}
\value{
numeric total seconds
}
\description{
Function to parse the time to create a total number of seconds
}
\keyword{internal}
