% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribeZoomAudio.R
\name{transcribeZoomAudio}
\alias{transcribeZoomAudio}
\title{Zoom Audio File Processing}
\usage{
transcribeZoomAudio(
  fileLocation,
  filePath,
  bucketName,
  jobName,
  numSpeakers,
  languageCode = "en"
)
}
\arguments{
\item{fileLocation}{character either "local" or "s3" - if local, then this function will upload the file to the specified bucket}

\item{filePath}{character    the path to the local file or to the s3 file (depending on whether it is "local" or "s3")}

\item{bucketName}{character    name of the s3 bucket where the finished transcript is stored}

\item{jobName}{character    name of the transcription job}

\item{numSpeakers}{numeric Specify how many speakers you expect (this helps AWS identify the speakers in the clip).}

\item{languageCode}{character language code. Default is "en".}
}
\value{
This does not return anything useful
}
\description{
This function starts an audio transcription job on AWS.
}
\details{
This function is useful for batch uploading audio files and starting transcription jobs for them.
This can be done with a local file (uploads to a specified s3 bucket) or with a file that already exists in an s3 bucket
}
\examples{

\dontrun{
transcribeZoomAudio(fileLocation="local",
bucketName="my-transcription-bucket",
filePath="mylocalfile.m4a",
jobName="mylocalfile.m4a",
languageCode="en-US")
}

}
