% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnTaking.R
\name{turnTaking}
\alias{turnTaking}
\title{Simple conversational turn-taking analysis}
\usage{
turnTaking(inputData, inputType, speakerId)
}
\arguments{
\item{inputData}{data.frame output from either processZoomChat or processZoomTranscript}

\item{inputType}{string of either 'chat' or 'transcript'}

\item{speakerId}{string giving the name of the variable with the identity of the speaker}
}
\value{
list of four data.frames. The named items provide different views of turntaking.
}
\description{
Generate a very basic analysis of the conversational turntaking in
either a Zoom transcript or a Zoom chat file.
}
\examples{
turn.out = turnTaking(inputData=sample_transcript_processed, 
inputType='transcript', 
speakerId='userName')

}
