% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggSentiment.R
\name{aggSentiment}
\alias{aggSentiment}
\title{Helper function to aggregate sentiment variables
This is just a simple helper function that is used to aggregate the
sentiment variables}
\usage{
aggSentiment(inputData, meetingId, indivId, sentMethod)
}
\arguments{
\item{inputData}{data.frame that has been output from textSentiment function}

\item{meetingId}{string that indicates the name of the variable containing the meeting ID}

\item{indivId}{string that indicates the name of the variable containing the speaker identity}

\item{sentMethod}{string that indicates what type of sentiment analysis to aggregate--must be either 'aws' or 'syuzhet'}
}
\value{
A list with data.frames. The first gives sentiment variables at the corpus level of analysis
The second gives sentiment variables at the speaker level of analysis.
}
\description{
Helper function to aggregate sentiment variables
This is just a simple helper function that is used to aggregate the
sentiment variables
}
\examples{
agg.out = aggSentiment(inputData=sample_transcript_sentiment_aws, 
meetingId="batchMeetingId", indivId = "userId", sentMethod="aws")

agg.out = aggSentiment(inputData=sample_transcript_sentiment_syu, 
meetingId="batchMeetingId", indivId = "userId", sentMethod="syuzhet")
}
