% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggSentiment.R
\name{aggSentiment}
\alias{aggSentiment}
\title{Helper function to aggregate sentiment variables
This is just a simple helper function that is used to aggregate the
sentiment variables}
\usage{
aggSentiment(inputData, speakerId)
}
\arguments{
\item{inputData}{data.frame that has been output from textSentiment function}

\item{speakerId}{string that indicates the name of the variable containing the speaker identity}
}
\value{
A list with data.frames. The first gives sentiment variables at the corpus level of analysis
The second gives sentiment variables at the speaker level of analysis.
}
\description{
Helper function to aggregate sentiment variables
This is just a simple helper function that is used to aggregate the
sentiment variables
}
\examples{
agg.out = aggSentiment(inputData=sample_transcript_sentiment_processed, speakerId="userName")
}
