% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchGrabVideoStills.R
\name{batchGrabVideoStills}
\alias{batchGrabVideoStills}
\title{Batch process video files, breaking them into stills}
\usage{
batchGrabVideoStills(batchInfo, sampleWindow)
}
\arguments{
\item{batchInfo}{the batchInfo data.frame that is output from batchProcessZoomOutput}

\item{sampleWindow}{an integer indicating how frequently you want to sample
images in number of seconds.}
}
\value{
a data.frame that gives information about the batch. Each record
corresponds to one video, with:
\itemize{
\item batchMeetingId - the meeting identifier
\item videoExists - boolean indicating whether the video file was there
\item sampleWindow - integer with the sampleWindow requested
\item numFramesExtracted - the number of image files that were saved
}
}
\description{
#' This helper calls grabVideoStills, which function currently
relies on either magick or ffmpeg to split a video file into images.
This function will save the images locally in the same
directory where the video is located.
}
\examples{
\dontrun{
vidBatchInfo = batchGrabVideoStills(batchInfo=zoomOut$batchInfo, 
sampleWindow=600)
}
}
