% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchProcessZoomOutput.R
\name{batchProcessZoomOutput}
\alias{batchProcessZoomOutput}
\title{Batch process files that have been downloaded from Zoom}
\usage{
batchProcessZoomOutput(batchInput, exportZoomRosetta = NULL)
}
\arguments{
\item{batchInput}{String giving the location of the xlsx file
that contains the information for the zoom meetings}

\item{exportZoomRosetta}{string giving the path for exporting the
zoomRosetta file to link up unique individual IDs manually}
}
\value{
a list that has a list item for each of the elements
of a Zoom output that are available--batchInfo, meetInfo, partInfo,
transcript, chat, and rosetta
}
\description{
Provide the location of a structured batchInput file and this
function will process a set of meetings at once.
}
\examples{
batchOut = batchProcessZoomOutput(batchInput=system.file('extdata', 
'myMeetingsBatch.xlsx', package = 'zoomGroupStats'))
}
