% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabVideoStills.R
\name{grabVideoStills}
\alias{grabVideoStills}
\title{Helper function to split a video into still frames
This function currently relies on a local install of ffmpeg to
split a video file into images. The images are saved locally
in a directory that can be specified by the user or the function.}
\usage{
grabVideoStills(inputVideo, sampleWindow, stillPath)
}
\arguments{
\item{inputVideo}{filepath to a video file}

\item{sampleWindow}{framerate for the pull of images}

\item{stillPath}{path to where the images should be saved}
}
\value{
does not return anything. just saves the images to a local directory
}
\description{
Helper function to split a video into still frames
This function currently relies on a local install of ffmpeg to
split a video file into images. The images are saved locally
in a directory that can be specified by the user or the function.
}
\examples{
\dontrun{
grabVideoStills(
inputVideo=system.file('extdata', 'sample_gallery_video.mp4', package = 'zoomGroupStats'), 
sampleWindow=45, stillPath="")
}
}
