% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabVideoStills.R
\name{grabVideoStills}
\alias{grabVideoStills}
\title{Helper function to split a video into still frames
This function currently relies on a local install of ffmpeg to
split a video file into images. The images are saved locally
in a directory that can be specified by the user or the function.}
\usage{
grabVideoStills(inputVideo, sampleWindow, tryffmpeg = TRUE)
}
\arguments{
\item{inputVideo}{full filepath to a video file}

\item{sampleWindow}{framerate in seconds for the pull of images}

\item{tryffmpeg}{boolean indicating whether you want it to try to use ffmpeg, which tends to be much faster than magick for splitting frames}
}
\value{
a vector containing the names of
the images that were saved
}
\description{
Helper function to split a video into still frames
This function currently relies on a local install of ffmpeg to
split a video file into images. The images are saved locally
in a directory that can be specified by the user or the function.
}
\examples{
\dontrun{
grabVideoStills(inputVideo='myMeeting.mp4', 
sampleWindow=45)
}
}
