% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importZoomRosetta.R
\name{importZoomRosetta}
\alias{importZoomRosetta}
\title{Import an edited zoomRosetta file that tells how to
link up zoom display names to some unique individual
identifier}
\usage{
importZoomRosetta(zoomOutput, zoomRosetta, meetingId)
}
\arguments{
\item{zoomOutput}{the output of batchProcessZoomOutput}

\item{zoomRosetta}{the path to an edited zoomRosetta}

\item{meetingId}{the name of the meetingId you want to use}
}
\value{
returns zoomOutput with the identifiers in zoomRosetta
merged to each of the list items if they are available
}
\description{
Import an edited zoomRosetta file that tells how to
link up zoom display names to some unique individual
identifier
}
\examples{
\dontrun{
importZoomRosetta(zoomOutput=batchOut, zoomRosetta="myEditedRosetta.xlsx", 
meetingId="batchMeetingId")
}
}
