% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTimeWindows.R
\name{makeTimeWindows}
\alias{makeTimeWindows}
\title{Time Windowing}
\usage{
makeTimeWindows(inputData, inputType, windowSize)
}
\arguments{
\item{inputData}{data.table Transcript data file to be analysed   #$MIKE$ - data.table? data.frame?}

\item{inputType}{character Type of data input}

\item{windowSize}{numeric Size of windows.}
}
\value{
A list with two objects.
\item{inputData}{data.frame the input data.}
\item{windows}{vector The window assignments}
}
\description{
Produces a set of time windows for a conversation.
}
\details{
This function takes in the transcript output of one of the other functions (either processZoomChat or processZoomTranscript).
}
