% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processZoomChat.R
\name{processZoomChat}
\alias{processZoomChat}
\title{Process a Zoom chat file
This function parses the data from the chatfile that is downloaded from the Zoom website.
NOTE: This is the file that accompanies a recording. This is not the file that you download directly within the window.
It is also not the one that is saved locally on your computer.
This is the file that you can access after a session if you record in the cloud.}
\usage{
processZoomChat(fname, sessionStartDateTime, languageCode)
}
\arguments{
\item{fname}{character}

\item{sessionStartDateTime}{character}

\item{languageCode}{character}
}
\value{
data.frame
}
\description{
Process a Zoom chat file
This function parses the data from the chatfile that is downloaded from the Zoom website.
NOTE: This is the file that accompanies a recording. This is not the file that you download directly within the window.
It is also not the one that is saved locally on your computer.
This is the file that you can access after a session if you record in the cloud.
}
\examples{
ch.out = processZoomChat(
fname=system.file('extdata', "sample_chat.txt", package = 'zoomGroupStats'), 
sessionStartDateTime = '2020-04-20 13:30:00', 
languageCode = 'en')
}
