% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processZoomChat.R
\name{processZoomChat}
\alias{processZoomChat}
\title{Zoom Chat File Processing}
\usage{
processZoomChat(fname, sessionStartDateTime, languageCode = "en")
}
\arguments{
\item{fname}{character the path to the local file where the chat file (txt) is saved.}

\item{sessionStartDateTime}{character the time that the actual session was launched. Format is YYYY-MM-DD HH:MM:SS}

\item{languageCode}{character language code. Default is "en".}
}
\value{
A data.frame containing a processed chat transcript with the following columns:
   \item{messageId}{an incremented numeric identifier for a message in the chat}
   \item{messageTime}{a timestamp for the message, based on the start of the Zoom session}
   \item{userName}{the name attached to the message}
   \item{message}{the text of the message}
   \item{messageLanguage}{the language code for the chat}
}
\description{
Processing the chat in parallel with a zoom call
}
\details{
This function parses the data from the chatfile that is downloaded from the Zoom website.
}
\examples{

\dontrun{
ch.out = processZoomChat(fname="~/Desktop/chat.txt", sessionStartDateTime="2020-04-01 17:56:34")
}      $MIKE$ - should probably default to read from internal character vector, with readLines option

}
