% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processZoomParticipantsInfo.R
\name{processZoomParticipantsInfo}
\alias{processZoomParticipantsInfo}
\title{Process participant information from a Zoom meeting export}
\usage{
processZoomParticipantsInfo(inputPath)
}
\arguments{
\item{inputPath}{character}
}
\value{
list of two data.frames with parsed information from the downloadable
Zoom participants file
\itemize{
\item meetInfo - provides the meeting level information that Zoom Cloud gives
\item partInfo - provides the participant level information that Zoom Cloud gives
}
}
\description{
This function parses the information from the downloadable meeting information file in Zooms reports section.
The function presumes that you have checked the box to
include the meeting information in the file.
That means that there is a header (2 rows) containing the zoom meeting information.
Following that header are four columns:
Name of user, user email, total duration, and guest.
}
\examples{
partInfo = processZoomParticipantsInfo(
system.file('extdata', "meeting001_participants.csv", package = 'zoomGroupStats')
)
}
