% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processZoomParticipantsInfo.R
\name{processZoomParticipantsInfo}
\alias{processZoomParticipantsInfo}
\title{Participants info parsing}
\usage{
processZoomParticipantsInfo(inputPath)
}
\arguments{
\item{inputPath}{character The path to the Zoom export file}
}
\value{
A list with two data.frame objects.
   \itemize{
     \item{by_chat} data.frame There is a single record containing the following info:
       \itemize{
         \item{meetingId}{Numeric meeting identifier from Zoom}
         \item{meetingTopic}{String meeting description}
         \item{meetingStartTime}{Start date-time of the meeting in YYYY-MM-DD HH:MM:SS}
         \item{meetingEndTime}{End date-time of the meeting in YYYY-MM-DD HH:MM:SS}
         \item{meetingDuration}{Number of minutes that the meeting lasted}
         \item{numParticipants}{Number of unique Zoom login names (original name) for this meeting}
       }
     \item{by_user} data.frame A set of measures aggregated to the level of the individual speaker (by unique login)
       \itemize{
         \item{userName}{String user name for this user}
         \item{userEmail}{String email address attached to this user's Zoom account}
         \item{userDuration}{Number of minutes that this user was logged in}
         \item{userGuest}{Indicator of whether this person is a guest user of Zoom}
       }
   }
}
\description{
Parses the information from the downloadable meeting information file in Zooms "reports" section.
}
\details{
The function presumes that you have
checked the box to include the meeting information in the file. That means
that there is a header (2 rows) containing the zoom meeting information.
Following that header are four columns:
Name (Original Name):		This is the name the user logged in first using
User Email: 				The user's email address (if logged in to Zoom)
Total Duration: 			How many minutes the user spent in the session
Guest:					Is this person a guest or not?
}
\examples{

\dontrun{
by_user = processZoomParticipantsInfo("~/Desktop/zoom_meeting_participants_export.csv")
}

}
