% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processZoomTranscript.R
\name{processZoomTranscript}
\alias{processZoomTranscript}
\title{Process Zoom transcript file}
\usage{
processZoomTranscript(fname, recordingStartDateTime, languageCode)
}
\arguments{
\item{fname}{String that is the path to the exported Zoom .vtt transcript chat file}

\item{recordingStartDateTime}{String that is the timestamp when the recording was started in YYYY-MM-DD HH:MM:SS}

\item{languageCode}{String denoting the language}
}
\value{
data.frame where each record is an utterance in the transcript
}
\description{
Process Zoom transcript file
}
\section{Zoom Recording Transcript File Processing}{
This function parses the data from the transcript file (.vtt) that is downloaded from the Zoom website.
NOTE: This is the file that accompanies a recording to the cloud.
}

\examples{
tr.out = processZoomTranscript(
fname=system.file('extdata', "sample_transcript.vtt", package = 'zoomGroupStats'),
recordingStartDateTime = '2020-04-20 13:30:00', 
languageCode = 'en')
}
