% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processZoomTranscript.R
\name{processZoomTranscript}
\alias{processZoomTranscript}
\title{Zoom Recording Transcript File Processing}
\usage{
processZoomTranscript(fileName, recordingStartDateTime, languageCode = "en")
}
\arguments{
\item{fileName}{character the path to the local file where the transcript file (vtt) is saved.}

\item{recordingStartDateTime}{character the date/time that the meeting recording started $MIKE$ format?}

\item{languageCode}{character language code. Default is "en".}
}
\value{
A list of analyses
   \item{utterance_id}{an incremented numeric identifier for a marked speech utterance}
   \item{utterance_start_seconds}{the number of seconds from the start of the recording (when it starts)}
   \item{utterance_start_time}{the timestamp for the start of the utterance}
   \item{utterance_end_seconds}{the number of seconds from the start of the recording (when it ends)}
   \item{utterance_end_time}{the timestamp for the end of the utterance}
   \item{utterance_time_window}{the number of seconds that the utterance took}
   \item{user_name}{the name attached to the utterance}
   \item{utterance_message}{the text of the utterance}
   \item{utterance_language}{the language code for the transcript}
}
\description{
This function parses the data from the transcript file (.vtt) that is downloaded from the Zoom website.
}
\details{
# Note: I plan to fix at a later point in time the timing issue. Specifically, it is not clear
where in Zoom's system I can get the actual time that the recording was started. This
is a problem for linking the transcript file up with the chat file.
One workaround for now (for research) would be to set recordings to auto-start. This is not ideal, though.
we should be able to know when the recording was started. It is embedded in the video, so could pull from there.
}
\examples{

\dontrun{
tr.out = processZoomTranscript(fileName="~/Desktop/transcript.vtt",
recordingStartDateTime="2020-04-01 17:56:34",
languageCode="en")
}

}
