% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_chat_processed}
\alias{sample_chat_processed}
\title{Parsed chat file in a 'Zoom' meeting}
\format{
A data frame with 30 rows of 9 variables:
\describe{
\item{batchMeetingId}{a character meeting identification variable}
\item{userName}{'Zoom' display name attached to this speaker}
\item{messageId}{an incremented numeric identifier for a marked chat message}
\item{messageSeconds}{when the message was posted as the number of seconds from the start of the recording}
\item{messageTime}{timestamp for message}
\item{message}{text of the message}
\item{messageLanguage}{language code of the message}
\item{userEmail}{character email address}
\item{userId}{numeric id of each speaker}
}
}
\source{
\url{http://zoomgroupstats.org/}
}
\usage{
sample_chat_processed
}
\description{
Parsed chat file in a 'Zoom' meeting
}
\keyword{datasets}
