% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textConversationAnalysis.R
\name{textConversationAnalysis}
\alias{textConversationAnalysis}
\title{Analyze conversation attributes
This function takes in the output of one of the other functions (either processZoomChat or processZoomTranscript)
and produces a set of conversation measures.}
\usage{
textConversationAnalysis(
  inputData,
  inputType,
  speakerId,
  doSentiment = FALSE,
  sentiDone = FALSE
)
}
\arguments{
\item{inputData}{data.frame}

\item{inputType}{character}

\item{speakerId}{character}

\item{doSentiment}{boolean}

\item{sentiDone}{boolean}
}
\value{
A list of two data.frames, with names conditional on your choice to analyze
a parsed transcript file or a parsed chat file. The first list item contains
statistics at the corpus level. The second list item contains statistics
at the speaker/messager level of analysis.
}
\description{
Analyze conversation attributes
This function takes in the output of one of the other functions (either processZoomChat or processZoomTranscript)
and produces a set of conversation measures.
}
\examples{
convo.out = textConversationAnalysis(inputData=sample_transcript_processed, 
inputType='transcript', speakerId='userName', doSentiment=FALSE, sentiDone=FALSE)
}
