% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textConversationAnalysis.R
\name{textConversationAnalysis}
\alias{textConversationAnalysis}
\title{Conversation Analysis Function}
\usage{
textConversationAnalysis(
  inputData,
  inputType,
  speakerId,
  doSentiment = FALSE,
  sentiDone = FALSE
)
}
\arguments{
\item{inputData}{data.table a transcript data file to be analysed   #$MIKE$ - data.table? data.frame?}

\item{inputType}{character type of data input}

\item{speakerId}{character name of column identifying speaker names}

\item{doSentiment}{boolean do you want to run a sentiment analysis on this input file? Default is FALSE}

\item{sentiDone}{boolean Does the file already have sentiment included? Default is FALSE}
}
\value{
A list with two data.frame objects.
   \itemize{
     \item{by_chat} data.frame A set of measures aggregated for the overall file (either chat or transcript).
       \itemize{
         \item{utterance_gap_x} Average number of seconds between one person's utterance and the next person's utterance.
         \item{utterance_gap_sd} The SD of the utterance gaps.
         \item{burstiness_raw}	This is a measure of how concentrated (in time) utterances are. It is not adjusted for # of utterances.
         \item{message_gap_x} The average number of seconds between one person's message and the next person's message.
         \item{message_gap_sd} The SD of the message gaps
         \item{} $MIKE I think there are some missing, yes?
         \item{}
       }
     \item{by_user} data.frame A set of measures aggregated to the level of the individual speaker (as identified by speakerId)
       \itemize{
         \item{utterance_gap_x} Average number of seconds (from the last utterance) that pass before this person makes an utterance.
         \item{message_gap_x} This is the average number of seconds (from the last message) that pass before this person sends a message.
         \item{} $MIKE I think there are some missing, yes?
         \item{}
       }
   }
}
\description{
Produces a set of conversation measures from a conversation transcript
}
\details{
This function takes in the transcript output of one of the other functions (either processZoomChat or processZoomTranscript).
}
\examples{

\dontrun{
textConversationAnalysis(inputData=outputOfOtherFunctions,
inputType="chat",
speakerID = "user_name",
doSentiment=TRUE,
sentiDone=FALSE)
}

}
