% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textSentiment.R
\name{textSentiment}
\alias{textSentiment}
\title{Conduct a sentiment analysis on text data}
\usage{
textSentiment(inputData, idVar, textVar, languageCodeVar)
}
\arguments{
\item{inputData}{data.frame that has been output by either the processZoomTranscript or processZoomChat functions}

\item{idVar}{name of variable that gives the unique identifier for this piece of text}

\item{textVar}{name of variable that contains the text}

\item{languageCodeVar}{name of variable that contains the language code}
}
\value{
returns the inputData with sentiment analysis variables appended
}
\description{
This function takes in the output of the chat and transcript functions. It then
conducts a sentiment analysis on an identified chunk of text
and returns the values.
To use this function, you must have an aws account that with privileges for the comprehend service
However you authenticate for AWS, you should do so before running the function.
}
\examples{
sent.out = textSentiment(inputData=sample_transcript_processed, idVar='utteranceId', 
textVar='utteranceMessage', languageCodeVar='utteranceLanguage')
}
