% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textSentiment.R
\name{textSentiment}
\alias{textSentiment}
\title{Text-based sentiment analysis function}
\usage{
textSentiment(inputData, idVar, textVar, languageCodeVar)
}
\arguments{
\item{inputData}{data.table a transcript data file to be analysed   #$MIKE$ - data.table? data.frame?}

\item{idVar}{character    the name of the id variable for the text}

\item{textVar}{character the name of the variable with the text in it}

\item{languageCodeVar}{character the variable containing the language code for each text chunk}
}
\value{
A data.frame including the inputData plus the following variables
   \item{sentClass}{the text-based sentiment classification of the message}
   \item{mixed}{the confidence level for the text being mixed sentiment}
   \item{negative}{the confidence level for the text being negative sentiment}
   \item{neutral}{the confidence level for the text being neutral sentiment}
   \item{positive}{the confidence level for the text being positive sentiment}
}
\description{
Conducts a sentiment analysis on the output of the chat/transcript functions.
}
\details{
To use this function, you must have an aws account with privileges for the comprehend
service. However you authenticate for AWS, you should do so before running the function.
}
\examples{

\dontrun{
sent.out = textSentiment(inputData=tr.out,
idVar = "utteranceId",
textVar = "utteranceMessage",
languageCodeVar = "utteranceLanguage")}
}
