% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnTaking.R
\name{turnTaking}
\alias{turnTaking}
\title{Turn-Taking Analysis}
\usage{
turnTaking(inputData, inputType = c("transcript"), speakerId)
}
\arguments{
\item{inputData}{data.table a transcript data file to be analysed   #$MIKE$ - data.table? data.frame?}

\item{inputType}{character type of data input}

\item{speakerId}{character name of column identifying speaker names}
}
\value{
A list of analyses
   \item{rawCount}{something}
   \item{rawCountNoDiag}{something}
   \item{pctByConvo}{something}
   \item{pctBySpeaker}{something}
   \item{dyadData}{something}
   \item{indivData}{something}
}
\description{
Conducts Turn-Taking Analysis on a transcript file.
}
\details{
Paragraph
}
\examples{

\dontrun{
trythis=turnTaking(data1,"transcript",speakerId="speakerId")
}
}
