% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowedConversationAnalysis.R
\name{windowedTextConversationAnalysis}
\alias{windowedTextConversationAnalysis}
\title{Window Analyses}
\usage{
windowedTextConversationAnalysis(
  inputData,
  inputType,
  speakerId,
  windowSize,
  doSentiment = FALSE,
  sentiDone = FALSE
)
}
\arguments{
\item{inputData}{data.table a transcript data file to be analysed   #$MIKE$ - data.table? data.frame?}

\item{inputType}{character type of data input}

\item{speakerId}{character name of column identifying speaker names}

\item{windowSize}{numeric Number of turns in each window of analysis.}

\item{doSentiment}{boolean do you want to run a sentiment analysis on this input file? Default is FALSE}

\item{sentiDone}{boolean Does the file already have sentiment included? Default is FALSE}
}
\value{
A list with two data.frame objects.
   \itemize{
     \item{by_chat} data.frame A set of measures aggregated for the overall file (either chat or transcript).
       \itemize{
         \item{}{}
         \item{}{}
         \item{}{}
       }
     \item{by_user} data.frame A set of measures aggregated to the level of the individual speaker (as identified by speakerId)
       \itemize{
         \item{}{}
         \item{}{}
       }
   }
}
\description{
Produces Windowed Conversation Anlaysis
}
\details{
Something else
}
\examples{

\dontrun{
audio.out = processZoomAudio(bucketName = "my-transcription-bucket",
jobName = "mylocalfile.m4a",
localDir = "path-to-local-directory-for-output",
speakerNames = c("Tom Smith", "Jamal Jones", "Jamika Jensen"),
recordingStartDateTime = "2020-06-20 17:00:00", writeTranscript=TRUE)
}


}
